extract trace CLI tests into sibling tests.rs
